requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
python-telegram-bot==13.8.1
requests==2.26.0